streamlit_succursale/requirements.txt
streamlit==1.49.1
supabase==2.4.6
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
Pillow==10.4.0
altair==5.3.0
